Support loading indicator in ControlsLabel
