+++ packages/scenes/src/variables/components/VariableValueSelect.tsx
@@ -1,39 +1,12 @@
 import { isArray } from 'lodash';
-import React, { useCallback } from 'react';
+import React from 'react';
 
-import { Icon, MultiSelect, Select, Tooltip } from '@grafana/ui';
-import { LoadingIndicatorProps } from 'react-select';
+import { MultiSelect, Select } from '@grafana/ui';
 
 import { SceneComponentProps } from '../../core/types';
 import { MultiValueVariable } from '../variants/MultiValueVariable';
-import { selectors } from '@grafana/e2e-selectors';
 import { VariableValue, VariableValueSingle } from '../types';
-
-
-const LoadingIndicator = ({ innerProps, ...props }: LoadingIndicatorProps & { selectProps: { onCancel: () => void }}) => {
-  const { onCancel } = props.selectProps;
-  const onMouseDown = useCallback(
-    (event: React.MouseEvent) => {
-      event.preventDefault();
-      event.stopPropagation();
-      onCancel();
-    },
-    [onCancel]
-  );
-
-  return (
-    <Tooltip content="Cancel query">
-      <Icon
-        className="spin-clockwise"
-        name="sync"
-        size="xs"
-        aria-label={selectors.components.LoadingIndicator.icon}
-        role="button"
-        onMouseDown={onMouseDown}
-      />
-    </Tooltip>
-  );
-};
+import { SelectLoadingIndicator } from '../../utils/LoadingIndicator';
 
 export function VariableValueSelect({ model }: SceneComponentProps<MultiValueVariable>) {
   const { value, key, loading } = model.useState();
@@ -51,8 +24,10 @@ export function VariableValueSelect({ model }: SceneComponentProps<MultiValueVar
       onChange={(newValue) => {
         model.changeValueTo(newValue.value!, newValue.label!);
       }}
-      onCancel={() => { model.cancel?.() }}
-      components={{ LoadingIndicator }}
+      onCancel={() => {
+        model.cancel?.();
+      }}
+      components={{ LoadingIndicator: SelectLoadingIndicator }}
     />
   );
 }
@@ -74,8 +49,10 @@ export function VariableValueSelectMulti({ model }: SceneComponentProps<MultiVal
       closeMenuOnSelect={false}
       isClearable={true}
       onOpenMenu={() => {}}
-      onCancel={() => { model.cancel?.() }}
-      components={{ LoadingIndicator }}
+      onCancel={() => {
+        model.cancel?.();
+      }}
+      components={{ LoadingIndicator: SelectLoadingIndicator }}
       onChange={(newValue) => {
         model.changeValueTo(
           newValue.map((v) => v.value!),
